Add China Support

+++ helpers/python.bat
@@ -11,7 +11,7 @@ if %errorlevel% neq 0 (
    goto :eof
 )
 
-
+echo ^> Done.
 echo Extracting Python...
 
 if exist "%python_extract_path%" (
@@ -25,6 +25,7 @@ if exist "%python_zip_path%" (
     del "%python_zip_path%"
 )
 
+echo ^> Done.
 echo Getting pip...
 
 powershell -Command "Invoke-WebRequest -Uri '%pip_url%' -OutFile '%pip_save_path%' -UseBasicParsing"
@@ -33,10 +34,12 @@ if %errorlevel% neq 0 (
    goto :eof
 )
 
+echo ^> Done.
 echo Installing pip...
 
 "%python_extract_path%\python.exe" "%pip_save_path%" >nul 2>&1
 
+echo ^> Done.
 echo Preparing python...
 
 if exist "%pth_file_path%" (
@@ -55,8 +58,25 @@ if exist "%PythonSavePath%" (
     del "%PythonSavePath%"
 )
 
-:: Install wheel, setuptools and poetry
-"%python_extract_path%\python.exe" -m pip install --no-warn-script-location wheel setuptools poetry
+:: Install wheel, setuptools and poetry in addition to the launcher requirements
+if %USE_TSINGHUA%%==1 (
+    "%python_extract_path%\python.exe" -m pip install --no-warn-script-location wheel setuptools poetry -i https://pypi.tuna.tsinghua.edu.cn/simple
 
-echo Done.
+    echo Downloading launcher requirements...
+    "%python_extract_path%\python.exe" -m pip install --no-warn-script-location psutil dearpygui GitPython -i https://pypi.tuna.tsinghua.edu.cn/simple
+
+    echo Installing DearPyGui-Markdown...
+    "%python_extract_path%\python.exe" -m pip install -e additional_modules/DearPyGui-Markdown-main -i https://pypi.tuna.tsinghua.edu.cn/simple
+    
+) else (
+    "%python_extract_path%\python.exe" -m pip install --no-warn-script-location wheel setuptools poetry
+
+    echo Downloading launcher requirements...
+    "%python_extract_path%\python.exe" -m pip install --no-warn-script-location psutil dearpygui GitPython
+
+    echo Installing DearPyGui-Markdown...
+    "%python_extract_path%\python.exe" -m pip install -e additional_modules/DearPyGui-Markdown-main
+)
+
+echo ^> Done.
 echo.
